Move gitprofile.config.js to root
