+++ src/components/blog/index.jsx
@@ -1,10 +1,10 @@
 import { getDevtoArticle, getMediumArticle } from 'article-api';
 import moment from 'moment';
 import { Fragment, useEffect, useState } from 'react';
-import config from '../../gitprofile.config';
 import { ga, skeleton } from '../../helpers/utils';
 import LazyImage from '../lazy-image';
 import PropTypes from 'prop-types';
+import config from '../../../gitprofile.config';
 
 const displaySection = () => {
   if (
