+++ src/components/head-tag-editor/index.jsx
@@ -1,8 +1,8 @@
 import { Fragment } from 'react';
 import { Helmet } from 'react-helmet-async';
 import PropTypes from 'prop-types';
-import config from '../../gitprofile.config';
 import { isThemeDarkish } from '../../helpers/utils';
+import config from '../../../gitprofile.config';
 
 const HeadTagEditor = ({ profile, theme }) => {
   return (
