+++ src/components/theme-changer/index.jsx
@@ -1,7 +1,7 @@
 import { AiOutlineControl } from 'react-icons/ai';
 import { skeleton } from '../../helpers/utils';
-import config from '../../gitprofile.config';
 import PropTypes from 'prop-types';
+import config from '../../../gitprofile.config';
 
 const ThemeChanger = ({ theme, setTheme, loading }) => {
   const changeTheme = (e, selectedTheme) => {
